Introduce predicates_enums.hpp
